backend/requirements.txt
google-genai==1.5.0
python-dotenv
pydantic
psycopg
fastapi[standard]